Create sig_img.php
Keith Wood's phenomenally easy to use jQuery signature pad plug in didn't have an easy way to convert the signatures, which are stored in a JSON string, to a more transportable image format.  This allows the user to capture the JSON string, which is basically a bunch of X & Y line coordinates, and draws the string onto an image on the fly.  This is my first contribution to GitHub -- still learning, so bear with me!